Move worker presence handler from generic_worker
